task_19_remake_in_3d
* весь функционал перенесен в 3d вид
