--- first_game/Assets/Prefabs/Brick 1.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c0ae7438215ad489aa1353e9ae268397
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
