+++ first_game/Assets/Scripts/Brick.cs
@@ -5,16 +5,21 @@ using UnityEngine;
 public class Brick : MonoBehaviour
 {
     private int hp = 1;
+    private int maxHP = 1;
 
     private float bottomScreen;
     private GameplayController gameController;
     private Vector3 newPosition;
     private bool move = false;
+    private float heightBrick;
+    private GameObject player;
 
     void Start()
     {
         gameController = Camera.main.GetComponent<GameplayController>();
         bottomScreen = Camera.main.ViewportToWorldPoint(new Vector3(0f, 0.01f, gameController.GetGameClipPlane())).y;
+        heightBrick = GetComponent<MeshRenderer>().bounds.extents.y * 2;
+        player = GameObject.FindGameObjectWithTag("Player");
 
         GameEvents.onDestroyAllBalls.AddListener(OnDestroyAllBalls);
         SwipeEvents.OnSwipeRight += OnSwipeRight;
@@ -36,6 +41,7 @@ public class Brick : MonoBehaviour
         if (collision.gameObject.CompareTag("Ball"))
         {
             hp--;
+            changeColor();
             if (hp <= 0)
                 Destroy(gameObject);
         }
@@ -43,7 +49,6 @@ public class Brick : MonoBehaviour
 
     private void MoveDown()
     {
-        float heightBrick = GetComponent<MeshRenderer>().bounds.extents.y * 2;
         transform.position -= new Vector3(0, heightBrick, 0);
         if (transform.position.y <= bottomScreen)
         {
@@ -52,9 +57,29 @@ public class Brick : MonoBehaviour
         }
     }
 
-    public void SetHP(int newHP) { hp = newHP; }
+    private void changeColor()
+    {
+        if (hp < maxHP && hp > 0)
+        {
+            Debug.Log("hp - " + hp);
+            Debug.Log("max - " + maxHP);
+            float percentHP = (float)hp / (float)maxHP;
+            Debug.Log("percent - " + percentHP);
+            Debug.Log("color - " + (1 - percentHP));
+            gameObject.GetComponent<MeshRenderer>().material.color = new Color(1, 0 + percentHP, 0 + percentHP, 1);
+        }
+    }
+
+    public void SetHP(int newHP, int newMaxHP)
+    {
+        hp = newHP;
+        maxHP = newMaxHP;
+        changeColor();
+    }
 
     public int GetHP() { return hp; }
+
+    public int GetMaxHP() { return maxHP; }
 
     private void OnDestroyAllBalls()
     {
@@ -63,7 +88,7 @@ public class Brick : MonoBehaviour
     
     private void OnSwipeRight()
     {
-        if (!move)
+        if (!move && player.GetComponent<PlayerCotrol>().GetCanShoot())
         {
             move = true;
             newPosition = BrickPosition.GetPositionOnRightSwipe(transform.position);
@@ -72,7 +97,7 @@ public class Brick : MonoBehaviour
 
     private void OnSwipeLeft()
     {
-        if (!move)
+        if (!move && player.GetComponent<PlayerCotrol>().GetCanShoot())
         {
             move = true;
             newPosition = BrickPosition.GetPositionOnLeftSwipe(transform.position);
@@ -83,173 +108,5 @@ public class Brick : MonoBehaviour
     {
         SwipeEvents.OnSwipeRight -= OnSwipeRight;
         SwipeEvents.OnSwipeLeft -= OnSwipeLeft;
-    }
-}
-
-public static class BrickPosition
-{
-    static readonly Dictionary<int, Vector3> front = new Dictionary<int, Vector3>
-    {
-        [1] = new Vector3(-3.03f, 5.5f, -3.03f),
-        [2] = new Vector3(-2.02f, 5.5f, -3.03f),
-        [3] = new Vector3(-1.01f, 5.5f, -3.03f),
-        [4] = new Vector3(0f, 5.5f, -3.03f),
-        [5] = new Vector3(1.01f, 5.5f, -3.03f),
-        [6] = new Vector3(2.02f, 5.5f, -3.03f)
-    };
-    static readonly Dictionary<int, Vector3> left = new Dictionary<int, Vector3>
-    {
-        [1] = new Vector3(-3.03f, 5.5f, 3.03f),
-        [2] = new Vector3(-3.03f, 5.5f, 2.02f),
-        [3] = new Vector3(-3.03f, 5.5f, 1.01f),
-        [4] = new Vector3(-3.03f, 5.5f, 0f),
-        [5] = new Vector3(-3.03f, 5.5f, -1.01f),
-        [6] = new Vector3(-3.03f, 5.5f, -2.02f)
-    };
-    static readonly Dictionary<int, Vector3> back = new Dictionary<int, Vector3>
-    {
-        [1] = new Vector3(3.03f, 5.5f, 3.03f),
-        [2] = new Vector3(2.02f, 5.5f, 3.03f),
-        [3] = new Vector3(1.01f, 5.5f, 3.03f),
-        [4] = new Vector3(0f, 5.5f, 3.03f),
-        [5] = new Vector3(-1.01f, 5.5f, 3.03f),
-        [6] = new Vector3(-2.02f, 5.5f, 3.03f)
-    };
-    static readonly Dictionary<int, Vector3> right = new Dictionary<int, Vector3>
-    {
-        [1] = new Vector3(3.03f, 5.5f, -3.03f),
-        [2] = new Vector3(3.03f, 5.5f, -2.02f),
-        [3] = new Vector3(3.03f, 5.5f, -1.01f),
-        [4] = new Vector3(3.03f, 5.5f, 0f),
-        [5] = new Vector3(3.03f, 5.5f, 1.01f),
-        [6] = new Vector3(3.03f, 5.5f, 2.02f)
-    };
-
-    static public Vector3 GetPositionOnRightSwipe(Vector3 pos)
-    {
-        if (IsEqualsPosition(pos, front[1]))
-            return left[1];
-        else if (IsEqualsPosition(pos, front[2]))
-            return left[2];
-        else if (IsEqualsPosition(pos, front[3]))
-            return left[3];
-        else if (IsEqualsPosition(pos, front[4]))
-            return left[4];
-        else if (IsEqualsPosition(pos, front[5]))
-            return left[5];
-        else if (IsEqualsPosition(pos, front[6]))
-            return left[6];
-        else if (IsEqualsPosition(pos, left[1]))
-            return back[1];
-        else if (IsEqualsPosition(pos, left[2]))
-            return back[2];
-        else if (IsEqualsPosition(pos, left[3]))
-            return back[3];
-        else if (IsEqualsPosition(pos, left[4]))
-            return back[4];
-        else if (IsEqualsPosition(pos, left[5]))
-            return back[5];
-        else if (IsEqualsPosition(pos, left[6]))
-            return back[6];
-        else if (IsEqualsPosition(pos, back[1]))
-            return right[1];
-        else if (IsEqualsPosition(pos, back[2]))
-            return right[2];
-        else if (IsEqualsPosition(pos, back[3]))
-            return right[3];
-        else if (IsEqualsPosition(pos, back[4]))
-            return right[4];
-        else if (IsEqualsPosition(pos, back[5]))
-            return right[5];
-        else if (IsEqualsPosition(pos, back[6]))
-            return right[6];
-        else if (IsEqualsPosition(pos, right[1]))
-            return front[1];
-        else if (IsEqualsPosition(pos, right[2]))
-            return front[2];
-        else if (IsEqualsPosition(pos, right[3]))
-            return front[3];
-        else if (IsEqualsPosition(pos, right[4]))
-            return front[4];
-        else if (IsEqualsPosition(pos, right[5]))
-            return front[5];
-        else if (IsEqualsPosition(pos, right[6]))
-            return front[6];
-        return Vector3.zero;
-    }
-
-    static public Vector3 GetPositionOnLeftSwipe(Vector3 pos)
-    {
-        if (IsEqualsPosition(pos, front[1]))
-            return right[1];
-        else if (IsEqualsPosition(pos, front[2]))
-            return right[2];
-        else if (IsEqualsPosition(pos, front[3]))
-            return right[3];
-        else if (IsEqualsPosition(pos, front[4]))
-            return right[4];
-        else if (IsEqualsPosition(pos, front[5]))
-            return right[5];
-        else if (IsEqualsPosition(pos, front[6]))
-            return right[6];
-        else if (IsEqualsPosition(pos, right[1]))
-            return back[1];
-        else if (IsEqualsPosition(pos, right[2]))
-            return back[2];
-        else if (IsEqualsPosition(pos, right[3]))
-            return back[3];
-        else if (IsEqualsPosition(pos, right[4]))
-            return back[4];
-        else if (IsEqualsPosition(pos, right[5]))
-            return back[5];
-        else if (IsEqualsPosition(pos, right[6]))
-            return back[6];
-        else if (IsEqualsPosition(pos, back[1]))
-            return left[1];
-        else if (IsEqualsPosition(pos, back[2]))
-            return left[2];
-        else if (IsEqualsPosition(pos, back[3]))
-            return left[3];
-        else if (IsEqualsPosition(pos, back[4]))
-            return left[4];
-        else if (IsEqualsPosition(pos, back[5]))
-            return left[5];
-        else if (IsEqualsPosition(pos, back[6]))
-            return left[6];
-        else if (IsEqualsPosition(pos, left[1]))
-            return front[1];
-        else if (IsEqualsPosition(pos, left[2]))
-            return front[2];
-        else if (IsEqualsPosition(pos, left[3]))
-            return front[3];
-        else if (IsEqualsPosition(pos, left[4]))
-            return front[4];
-        else if (IsEqualsPosition(pos, left[5]))
-            return front[5];
-        else if (IsEqualsPosition(pos, left[6]))
-            return front[6];
-        return Vector3.zero;
-    }
-
-    static public bool IsEqualsPosition(Vector3 pos1, Vector3 pos2)
-    {
-        return (pos1.x == pos2.x && pos1.z == pos2.z);
-    }
-
-    static public Vector3 GetBrickCoord(int side, int position)
-    {
-        switch (side)
-        {
-            case Constants.FRONT_SIDE:
-                return front[position];
-            case Constants.LEFT_SIDE:
-                return left[position];
-            case Constants.BACK_SIDE:
-                return back[position];
-            case Constants.RIGHT_SIDE:
-                return right[position];
-            default:
-                return Vector3.zero;
-        }
     }
 }
