+++ first_game/Assets/Scripts/BrickPosition.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b35edf409899a450287ebb86eda62b96
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
