+++ first_game/Assets/Scripts/GameEvents.cs
@@ -8,5 +8,6 @@ public static class GameEvents
 
     public static UnityEvent onDestroyAllBalls = new UnityEvent();
     public static UnityEvent onBrickTouchingBottom = new UnityEvent();
+    public static UnityEvent onClickLoosingMessage = new UnityEvent();
 
 }
