+++ first_game/Assets/Scripts/Swipe Controller/SwipeEvents.cs
@@ -29,8 +29,11 @@ public class SwipeEvents : MonoBehaviour
 
     void Update()
     {
-        PreventMultipleTriggersSimultaneously();
-        SwipeInputs();
+        if (Time.timeScale > 0)
+        {
+            PreventMultipleTriggersSimultaneously();
+            SwipeInputs();
+        }
     }
 
     #endregion
